% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rss_distribution.R
\name{plot_rss_distribution}
\alias{plot_rss_distribution}
\title{Plot RSS Distribution per Sample}
\usage{
plot_rss_distribution(result_df)
}
\arguments{
\item{result_df}{A data frame containing "Sample" and "RSS" columns (output of calculate_rss()).}
}
\value{
A ggplot2 object.
}
\description{
This function generates a bar plot showing the RSS score distribution across all samples.
}
